+++ R/plotting.R
@@ -644,7 +644,7 @@ my_dens <- function(data, mapping, center_point = 0, ...) {
 get_label_centers <- function(obj, group_var = "clusters", reduced_dim = "UMAP") {
   # To any scatter of a umap can add + geom_label_repel(data = get_label_centers(sce), aes(x = x, y = y, label = clusters))
 
-  if ("SingleCellExperiment" %in% class(obj)) {
+  if (class(obj) %in% c("SingleCellExperiment", "Milo")) {
     x_means <- lapply(split(reducedDim(obj, reduced_dim)[, 1], obj[[group_var]]), mean) %>% unlist()
     y_means <- lapply(split(reducedDim(obj, reduced_dim)[, 2], obj[[group_var]]), mean) %>% unlist()
     centers <- data.frame(x = x_means, y = y_means)