+++ NEWS.md
@@ -1,3 +1,5 @@
+# scatools (development version)
+
 # scatools 0.1.1
 
 * Fixed H5AD conversion