+++ NEWS.md
@@ -1,7 +1,9 @@
 # scatools (development version)
 
-* Added `segment=FALSE` as default
 * Replace `add_gc_cor`, `segment_cnv`, and `merge_segments` parallel backend to use `BiocParallel`
+* Added `segment=FALSE` as default
+* Minor bugfixes
+
 
 # scatools 0.1.1