+++ NEWS.md
@@ -1,4 +1,4 @@
-# scatools (development version)
+# scatools 0.1.2
 
 * Replace `add_gc_cor`, `segment_cnv`, and `merge_segments` parallel backend to use `BiocParallel`
 * Added `segment=FALSE` as default